requirements.txt
pygame>=2.1.0
tkinter
python-dotenv>=0.19.0
pytest>=6.2.5
pillow>=8.3.1
socketio>=0.2.1
flask>=2.0.1
flask-socketio>=5.1.1
numpy>=1.21.0